+++ src/app/core/models/db/category-mapping.model.ts
@@ -1,11 +1,12 @@
-import { MappingDestination } from "./mapping-destination.model";
-import { MappingSource } from "./mapping-source.model";
+
+import { DestinationAttributeDetail } from "./destination-attribute.model";
+import { ExpenseAttribute } from "./expense-attribute.model";
 import { MappingResult } from "./mapping.model";
 
 export type CategoryMappingPost = {
-  source_category: MappingSource;
-  destination_account?: MappingDestination;
-  destination_expense_head?: MappingDestination;
+  source_category: ExpenseAttribute;
+  destination_account?: DestinationAttributeDetail;
+  destination_expense_head?: DestinationAttributeDetail;
   created_at?: Date;
   updated_at?: Date;
   workspace: number;
