--- src/app/core/models/db/destination-attribute-detail.model.ts
@@ -1,9 +1,0 @@
-export type DestinationAttributeDetail = {
-    email: string;
-    customer_id: string;
-    customer_name: string;
-    gl_account_no: string;
-    gl_account_title: string;
-    is_reimbursable: boolean;
-    country: string;
-  };
