+++ src/app/core/models/db/destination-attribute.model.ts
@@ -1,3 +1,17 @@
+export type DefaultDestinationAttribute = {
+  id: string,
+  name: string,
+};
+
+export type DestinationAttributeDetail = {
+  email: string;
+  customer_id: string;
+  customer_name: string;
+  gl_account_no: string;
+  gl_account_title: string;
+  is_reimbursable: boolean;
+  country: string;
+};
 
 export type DestinationAttribute = {
   id: number;
@@ -9,10 +23,5 @@ export type DestinationAttribute = {
   created_at: Date;
   updated_at: Date;
   workspace: number;
-
-};
-
-export type DefaultDestinationAttribute = {
-  id: string,
-  name: string,
+  detail?: DestinationAttributeDetail;
 };
