--- src/app/core/models/db/mapping-destination.model.ts
@@ -1,14 +1,0 @@
-import { DestinationAttributeDetail } from "./destination-attribute-detail.model";
-
-export type MappingDestination = {
-    id?: number | null;
-    attribute_type?: string;
-    display_name?: string;
-    value?: string;
-    destination_id?: string;
-    active?: boolean;
-    created_at?: Date;
-    updated_at?: Date;
-    workspace?: number;
-    detail?: DestinationAttributeDetail;
-  };
