+++ src/app/core/models/db/mapping.model.ts
@@ -13,14 +13,14 @@ export type FyleField = {
 
 export type MappingResult = {
     id: number;
-    attribute_type?: string;
-    display_name?: string;
-    value?: string;
-    source_id?: number;
-    auto_mapped?: boolean;
-    active?: boolean;
-    created_at?: Date;
-    updated_at?: Date;
-    workspace?: number;
+    attribute_type: string;
+    display_name: string;
+    value: string;
+    source_id: number;
+    auto_mapped: boolean;
+    active: boolean;
+    created_at: Date;
+    updated_at: Date;
+    workspace: number;
     detail?: ExpenseAttributeDetail;
   };
